requirements.dev.txt
isort
flake8
mypy
django-crispy-forms
crispy-bootstrap5